Relocate mock file
